Redesign event edit page
Fix bug from react-final-form migration normalize->parse
Fix bug in ImageUpload filetypes
