+++ app/components/Form/ImageUploadField.css
@@ -1,3 +1,5 @@
+@import url('~app/styles/variables.css');
+
 .base {
   position: relative;
   overflow: hidden;
@@ -5,6 +7,6 @@
 
 .coverImage {
   width: 100%;
-  height: 300px;
+  aspect-ratio: 3.3;
   background: var(--additive-background);
 }
