+++ app/components/Upload/UploadImage.css
@@ -9,7 +9,7 @@
 
 .dropArea {
   border-radius: inherit;
-  height: inherit;
+  aspect-ratio: inherit;
   width: inherit;
   z-index: 0;
   position: absolute;
@@ -29,13 +29,17 @@
 .placeholderContainer {
   overflow: hidden;
   border-radius: inherit;
-  height: inherit;
+  aspect-ratio: inherit;
   width: inherit;
   position: absolute;
   background: rgba(var(--rgb-max), var(--rgb-max), var(--rgb-max), 50%);
   font-weight: 500;
   padding: 10px 40px;
   text-align: center;
+
+  @media (--small-viewport) {
+    line-height: 1.3;
+  }
 }
 
 .inModalUpload .placeholderContainer {
