+++ app/routes/admin/email/components/EmailListEditor.tsx
@@ -156,7 +156,7 @@ const EmailListEditor = () => {
             name="requireInternalAddress"
             type="checkbox"
             component={CheckBox.Field}
-            normalize={(v) => !!v}
+            parse={(v) => !!v}
           />
 
           <Field
