+++ app/routes/companyInterest/components/CompanyInterestPage.tsx
@@ -85,7 +85,7 @@ const SemesterBox = ({
         label={semesterToText({ ...fields.value[index], language })}
         type="checkbox"
         component={CheckBox.Field}
-        normalize={(v) => !!v}
+        parse={(v) => !!v}
       />
     ))}
   </Flex>
@@ -106,7 +106,7 @@ const SurveyOffersBox = ({
         label={SURVEY_OFFERS[surveyOffersToString(item)][language]}
         type="checkbox"
         component={CheckBox.Field}
-        normalize={(v) => !!v}
+        parse={(v) => !!v}
       />
     ))}
   </Flex>
@@ -138,7 +138,7 @@ const EventBox = ({
               label={EVENTS[eventToString(key)][language]}
               type="checkbox"
               component={CheckBox.Field}
-              normalize={(v) => !!v}
+              parse={(v) => !!v}
             />
           ))}
         </Flex>
@@ -162,7 +162,7 @@ const TargetGradeBox = ({
         label={TARGET_GRADES[targetGradeToString(key)][language]}
         type="checkbox"
         component={CheckBox.Field}
-        normalize={(v) => !!v}
+        parse={(v) => !!v}
       />
     ))}
   </Flex>
@@ -183,7 +183,7 @@ const OtherBox = ({
         label={readmeIfy(OTHER_OFFERS[otherOffersToString(key)][language])}
         type="checkbox"
         component={CheckBox.Field}
-        normalize={(v) => !!v}
+        parse={(v) => !!v}
       />
     ))}
   </Flex>
@@ -204,7 +204,7 @@ const CollaborationBox = ({
         label={COLLABORATION_TYPES[collaborationToString(key)][language]}
         type="checkbox"
         component={CheckBox.Field}
-        normalize={(v) => !!v}
+        parse={(v) => !!v}
       />
     ))}
   </Flex>
