+++ app/routes/surveys/components/SurveyEditor/Question.tsx
@@ -150,7 +150,7 @@ const Question = ({
             label="Obligatorisk"
             type="checkbox"
             component={CheckBox.Field}
-            normalize={(v) => !!v}
+            parse={(v) => !!v}
           />
 
           <ConfirmModal
