+++ app/styles/overlay.css
@@ -6,7 +6,7 @@
   border-radius: var(--border-radius-lg);
   position: absolute;
   margin-top: 10px;
-  z-index: 2;
+  z-index: 70;
 
   @media (--small-viewport) {
     width: calc(100% - 10px);
