+++ cypress/support/utils.js
@@ -24,7 +24,7 @@ export const selectField = (name) =>
     .parent()
     .parent();
 export const selectFieldDropdown = (name) =>
-  selectField(name).find(`[id=react-select-${name}-listbox]`);
+  cy.get(`[id=react-select-${name}-listbox]`);
 
 export const selectFromSelectField = (name, option, search) => {
   selectField(name).click();
