+++ packages/lego-bricks/src/components/Modal/Modal.module.css
@@ -13,7 +13,7 @@
   background: rgba(var(--rgb-min), var(--rgb-min), var(--rgb-min), 10%);
   backdrop-filter: blur(3px);
   inset: 0;
-  z-index: 3;
+  z-index: 90;
   animation: fade-in var(--easing-medium) forwards;
 }
 
@@ -46,7 +46,7 @@ html[data-theme='dark'] .backdrop {
   background: var(--color-white);
   outline: none;
   overflow-y: auto;
-  z-index: 4;
+  z-index: 100;
   box-shadow: var(--shadow-sm);
   border-radius: var(--border-radius-lg);
   animation: pop-in var(--easing-medium) forwards;
